Center the fields when the window gets bigger
Just import the Tkinter and tkMessageBox modules in gui.py, I doubt
that anyone would really want a library to cause a quit.  Also since I
rewrote the GUI class to inherit from Frame the check and message
couldn't be kept in the GUI.__init__ anymore.

Since the "old way" doesn't seem to let you specify how to grid the
"master" property, it seemed important to rewrite the class to inherit
from Frame so that we can.

Move the starting of the mainloop to the gdispass script.  I think
that only crazy OOP-only languages occasionally have constructors
cause side-effects, although hopefully not too much.