Add new algorithm Dispass2
Dispass2 is a new algorithm introduced to solve a problem where
passwords get expired. Instead of altering the label used for creating
the passphrase, a sequence number can be optionally specified.

Dispass1 passphrases are incompatible with Dispass2 passphrases,
manually altering the passphrase on the target login instance is needed
for each label when changing to this new algorithm.

Thanks to Tom (ryuslash) Willemsen for coming up with this idea.