Use global settings object for default values
Define a new 'Settings' class for initializing default values for things
like passphrase_length, algorithm and sequence_number. Create an
instance of Settings at the earliest points, passing it around to the
constructors of each component.

This makes it possible to add support for initializing values from a
settings file, only having to parse it once.