Change default file location to XDG_CONFIG_HOME
A labelfile is a data file, but it is meant to be kept private.
Therefore moving it to a subdir in .config seems to be a better option.

IMPORTANT:
The default file location for the labelfile on *nix is now
'$HOME/.config/dispass/labels', if you are using the development
version you will have to move it to this location manually or use the
DISPASS_LABELFILE environment variable.