Add "add" option to dispass-label
With the `add' option a new label can be added to the labelfile in a
non-interactive way.  It uses a "labelspec" to determine each
parameter for the label, which looks like:

    <label>[:<length>[:<algorithm>[:<sequence number>]]]

Where only the label is actually required, each part is only necessary
when the next is used, but each part may also be emtpy, here are some
examples:

    test

Use the name "test" and the default values for everything else.

    test:15

Use the name "test" with a length of 15 and the default values for
everything else.

    test:15:dispass2:3

Use the name "test" with a length of 15, algorithm "dispass2" and
sequence number 3 and the default for everything else (currently there
is nothing else).

    test::dispass2:3

Use the name "test" with the default length, algorithm "dispass2" and
sequence number 3 and the default for everything else.

Downsides include: this excludes `:' from being used in labels here,
and requires knowledge of the order of parameters.