Improve quitting and clearing of interactive edit
When `^C' is received by the interactive editor, just return to the
prompt.  This way someone could decide not to add or remove some label
without being kicked out of the entire program.  This also makes
clearing lines very easy.

To exit a `^D' (EOF) can be sent and it won't show a traceback
anymore.