Add generate shortcut in dispass add subcommand
Add `-g --generate` flags for quickly generating the passphrase after
adding it to the labelfile.

Before this change you would generate a passphrase for the first time,
like this:

    $ dispass add someaccountname:15:dispass2
    $ dispass generate -v someaccountname

Now, you can accomplish the same with just a single command:

    $ dispass add -g someaccountname:15:dispass2

When using the `-g --generate` flag, the passphrase will always be
checked twice and curses is always used if it is found. The first
mentioned method is of course still valid. This isn't a substitute of
the generate subcommand, it's just a shortcut.