Add update command
Add a new `update' command to make changing settings in the label file a
little easier.

The syntax for specifying changes is based on the syntax used by the add
command and consists of `:'-separated fields being specified in the
order of size, algorithm, sequence number. Each item is optional but the
colons are not, so to only change the sequence number to `4' for the
label `test' one would use:

dispass update test ::4

This would have no effect if the `test' label used the `dispass1'
algorithm.