Add disabled information to labelfile
Indicates whether a label is disabled in the labelfile. This indicates
whether it should be considered for printing under normal circumstances
(not explicitly requesting all labels) for printing and such.

Disabling a label allows the user to keep certain labels in the
labelfile without being bothered by outdated labels in listing output
and the like.

Represent the disabled state in the list header with "X".