Fix python3 dict_keys object in label Combobox
In python 3 dict.keys() returns an iterable instead of a list. When no
filtering is done because a user clicks the dropdown of the Combobox
without typing a few chars first, the Combobox is populated with the
textual representation of the python object instead of a nicely
duck-typed list.

E.g.: dict_keys(['Google', 'Hotmail', 'Facebook'])

Fix this by casting to a list. Because this operation is not atomic, copy the
dict beforehand to prevent any (rare) exceptions.