Fix labelfile interaction when requesting help
When a user (or testing framework) requests help/usage information for
a subcommand that deals with the labelfile when no labelfile has been
initialized yet, he/she is not presented with the help information.
Instead, the user is prompted to create a new labelfile interactively.

This commit changes the decorators read_labels and write_labels to
skip labelfile checks when help is requested via the `help` flags of a
subcommand. This way, the help message is always shown, regardless of
the contents or location of a labelfile.