Use automatic script creation instead of scripts
Setuptools can automatically generate scripts to match local conventions
on both Windows and POSIX platforms.

https://setuptools.readthedocs.io/en/latest/setuptools.html#automatic-script-creation

Previously the bin/* scripts correctly run the program, returning exit
status code to the shell using `sys.exit(...)`. Setuptools does this for
us now, so we need a function or method that returns the exit status
code, but does not call `sys.exit`.

This commit adds a new `dispass.main` module with two functions,
and configures setup.py to use them for creating the `dispass` and
`gdispass` scripts. The bin/* scripts are removed.